+++ scout/server/blueprints/cases/views.py
@@ -5,7 +5,7 @@ import datetime
 from operator import itemgetter
 
 from flask import (abort, Blueprint, current_app, redirect, render_template,
-                   request, url_for, send_from_directory, jsonify)
+                   request, url_for, send_from_directory, jsonify, flash)
 from flask_login import current_user
 from flask_weasyprint import HTML, render_pdf
 from dateutil.parser import parse as parse_date
@@ -404,3 +404,14 @@ def default_panels(institute_id, case_name):
     panel_ids = request.form.getlist('panel_ids')
     controllers.update_default_panels(store, current_user, institute_id, case_name, panel_ids)
     return redirect(request.referrer)
+
+
+@cases_bp.route('/<institute_id>/<case_name>/multiqc')
+def multiqc(institute_id, case_name):
+    """Load multiqc report for the case."""
+    data = controllers.multiqc(store, institute_id, case_name)
+    if data['case'].get('multiqc') is None:
+        return abort(404)
+    out_dir = os.path.abspath(os.path.dirname(data['case']['multiqc']))
+    filename = os.path.basename(data['case']['multiqc'])
+    return send_from_directory(out_dir, filename)